// ui/src/lib/gen/RepoDraft.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CodeHostBinding } from "./CodeHostBinding";

export type RepoDraft = { id: string, path: string, primary: boolean, remote: string, base: string, remote_url: string | null, code_host: CodeHostBinding | null, };